Initial Commit of the OASIS API that is the core of Our World and the OASIS (Open Advanced Sensory Imersion System) System. This is where the core Avatar/Profile is stored and also controls the Karma System. It also provides HoloOASIS SDK/Lib which connects .NET to Holochain. This will be gifted forward to the Holochain community along with the HoloUnity3D SDK/Lib coming soon... ;-)
So far some basic interfaces have been defined such as IOASISStorage & IOASISNetwork, which HoloOASIS will implement. This check in also included the first version of the HoloNETClient along with a Test Harness to test it. Unit tests coming later along with a Zome Proxy Generator, .NET HDK and so much more! ;-)

This also includes the bare bones of the Our World holochain app. Much more to come soon... :)