- STAR ODK - Integrated Avatar Authentication for BeamIn method. (Calls same new authentication method the OASIS API REST Servive uses.)
- Moved avatar authentication/security methods from OASIS API REST WebAPI to OASIS.API.Core.AvatarManager so STAR, Unity (Our World) & others can share the authentication methods.

Yes, they could just call the OASIS API REST Service but that would have an additional overhead with the HTTP Request. This way they have a choice of using that or this new fully native integrated version.

The bonus of this new way is that even if they are offline they can still log-in if they are using a local OASIS Storage Provider such as HoloOASIS, etc.... :)

Feels good to FINALLY return to STAR ODK! ;-)