WebUI.HTML cleanup and project sub-folder creation
All files that are no longer relevant and/or redundant have been removed.
Each framework now has it's own directory. To reduce confusion as much as possible, a temporary README has been created in the root directory, outlining the new structure, and will be updated and added to as needed.