Proposed Documentation
This adds in a sample pull request template to be used for automating the approval or acceptance of future pull requests. If this format is not followed or information is missing (the Jira task specifically) , then the pull request will be rejected until corrected.

Also included is the style guide for React/JSX as well as instructions for locally installing it to lint your code for you as you go, which is optional.

Last doc is the documentation for utilizing Smart Commits, which will move your Jira task through the different stages based on key words used in your commit messages. We should be used to this sort of structure as it is very similar to the daily standup we've been doing recently.

Any and all changes suggestions are welcome, feel free to collaborate on it and add or remove whatever isn't relevant from the style guide. (install instructions and Smart Commit documentation obviously excluded)